Generalize heap permutation
New traits were added to represent types that can be permuted and
types that can permute. This should help with the expansion of the
crate to other permutation and combination algorithms in the
future.

The ability to permute strings and based on either codepoints or
graphemes was also added in. It should be noted that codepoint
based permutation hasn't been tested as of this commit, so it
can likely produce invalid utf8 sequences. Use caution when
permuting non ascii data by codepoint. This drawback can be the
subject of a future commit.